Adopt a Delta based party mode
The old party mode implementation would sync the complete state of the queue over
every time something changed, this often let to bugs and wasnt a great approach.
With this change we adopt a delta based model, where we are only syncing what has changed.
We do still send over the whole climb when adding to the queue to avoid having to make db queries